+ Improved VST preset navigation in VST window. Known problem: doesn't update if preset is changed from VSTs own preset navigation. + Improvements in IT compatibility playmode. When loading ITs, if file doesn't seem modplug made, then automatically setting IT compatibility playmode. The playmode setting can be toggled in song properties dialog, and the autodetection with an ini setting. + 'key binding not understood'-messages can be disabled with an ini setting. . Fixed wrong file version in IT compatibility save. (http://lpchip.com/modplug/viewtopic.php?t=1723) . Fixed possible unnotified overwriting of file occurring for example when dealing with a new file. (http://lpchip.com/modplug/viewtopic.php?t=1813) . Fixed broken note preview for certain type of instruments in instrument tab. (http://lpchip.com/modplug/viewtopic.php?t=1996) . Fixed possible crash on channel mode wav export when a channel has a long name. (http://lpchip.com/modplug/viewtopic.php?t=2005) . Fixed pattern position indicator jump when clicking VST-box in channel header. (http://lpchip.com/modplug/viewtopic.php?t=2014) . CMainFrame::m_csExecutablePath renamed and should now contain the path of executable directory, not that of the working directory. / Tuning related changes/improvements
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@192 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/OrderToPatternTable.cpp
@@ -9,7 +9,7 @@ DWORD COrderToPatternTable::UnSerialize(const BYTE* const src, const int memLeng
 {
 	if(memLength < sizeof(s_Version) + sizeof(size_t)) return 0;
 	WORD version = 0;
-	size_t s = 0;
+	uint32 s = 0;
 	DWORD memPos = 0;
 	memcpy(&version, src, sizeof(version));
 	memPos += sizeof(version);
@@ -55,7 +55,7 @@ bool COrderToPatternTable::Serialize(FILE* f) const
 
 	//Version
 	fwrite(&s_Version, sizeof(s_Version), 1, f);
-	const size_t s = size();
+	const uint32 s = size();
 	//Size
 	fwrite(&s, sizeof(s), 1, f);
 	//Values
